file&move: UITextField+Ext
